fix(profiles): govern public profile scores with score_visibility
The public user and team profiles handed every player's score and full solve
history to any verified account regardless of score_visibility. GET /users/{id}
and GET /teams/{id} gate only on account visibility, and their handlers emitted
the score and solve list raw — the L4 score-redaction path (Redactor.Account)
had no caller, so the class gate was the only wall and it does not watch scores.

score_visibility is applied here by redaction, not by a route gate: account
visibility already decides whether the page exists, so gating scores at the route
would turn "scores are hidden" into a 404 and hide an account that is meant to be
visible. Instead the score becomes null on the wire (never 0) and the solve
history is withheld as omission — its length is a solve count, so serving rows
beside a nulled score would hand that number back by difference. The team page is
the same shape: the total and every member's contribution are withheld while the
roster, which is account data, stays. Admins and the own-team view see full
detail; the freeze horizon still clamps the history exactly as before.

This wires Redactor.Account (previously dead) into the real path and adds
ProfileSolveList / TeamMember beside the existing ChallengeSolveList, so there is
one redaction mechanism, not two. The embedded solve list gains a LIMIT — a
public document must not be an unbounded scrape or a slow query — and the score
total, summed independently, is unaffected by the cap; the team roster gains a
defensive LIMIT atop the enrollment-time team_size cap.

The score-over-time endpoint (GET /scoreboard/{id}) already gates on VisScore at
the route and is left as is; TestS54 pins that it stays a 404, not a redaction.

+++ internal/domain/policy/named_policies_test.go
@@ -351,7 +351,7 @@ func TestOwnScoreLiveUnderFreezeAndHiddenScores(t *testing.T) {
 		t.Fatal("score_visibility=hidden means scores are not visible, even to build the redactor")
 	}
 
-	score, place := 1337, 4
+	score, place := int64(1337), 4
 	f := policy.AccountFields{Score: &score, Place: &place}
 	r.Self(&f)
